DWORD64 Again

+++ mod/src/mod.cpp
@@ -457,8 +457,8 @@ DWORD WINAPI ThreadProc(LPVOID lpParam) {
                     }
 
                     // Purchased Location // TODO: test
-                    if (a & 0b0100) game.extraPurchased |= (((DWORD64)1) << (i+((DWORD)1)));
-                    else game.extraPurchased &= ~(DWORD64)(((DWORD64)1) << (i+((DWORD)1)));
+                    if (a & 0b0100) game.extraPurchased |= (((DWORD64)1) << (DWORD64)(i+((DWORD)1)));
+                    else game.extraPurchased &= ~(DWORD64)(((DWORD64)1) << (DWORD64)(i+((DWORD)1)));
                     game.extraPurchasedPrev = game.extraPurchased;
                 }
 
@@ -480,8 +480,8 @@ DWORD WINAPI ThreadProc(LPVOID lpParam) {
                         else *game.levels.levelRedBrick[i] = 0;
                     }
                     // Purchased Item // TODO: test
-                    if (a & 0b1000) game.extraPurchased |= (((DWORD64)1) << (i+((DWORD)1)));
-                    else game.extraPurchased &= ~(DWORD64)(((DWORD64)1) << (i+((DWORD)1)));
+                    if (a & 0b1000) game.extraPurchased |= (((DWORD64)1) << (DWORD64)(i+((DWORD)1)));
+                    else game.extraPurchased &= ~(DWORD64)(((DWORD64)1) << (DWORD64)(i+((DWORD)1)));
                     game.extraPurchasedPrev = game.extraPurchased;
                 }
                 
@@ -628,7 +628,7 @@ DWORD WINAPI ThreadProc(LPVOID lpParam) {
                 // TODO: test
                 std::cout << "Red Brick Unlocked!" << std::endl;
                 game.powerBrickState[i-515] |= 0b1000;
-                if (!game.isInShop()) game.extraPurchased |= (1 << (i-515+1));
+                if (!game.isInShop()) game.extraPurchased |= (((DWORD64)1) << (DWORD64)(i-((DWORD64)515)+((DWORD)1)));
